Improve HDBSCAN implementation + Use incremental processing after clustering
